Added first Arduino test

+++ randombytes.h
@@ -1,1 +1,1 @@
-void randombytes(unsigned char * ptr,unsigned int length);
+void randombytes(unsigned char * ptr,unsigned long long length);
